Set up Shellcheck as a dedicated backend
