+++ pants-plugins/examples/bash/lint/shellcheck/BUILD
@@ -1,8 +1,5 @@
 # Copyright 2020 Pants project contributors.
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from examples.bash.target_types import BashTarget
-
-
-def target_types():
-    return [BashTarget]
+# Pants will infer a dependency on `pants-plugins:pantsbuild.pants`.
+python_library()
